# service/s3: Use interfaces assertions instead of ValuesAtPath for S3 field lookups.