… service/s3/api.go