Improves the performance across the board for all S3 API calls by removing the usage of `ValuesAtPath` being used for every S3 API call.

With a simple S3 HeadObject call the following benchmarks improved the performance of the API call significantly.

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkS3HeadObject-4     190419        169992        -10.73%

benchmark                   old allocs     new allocs     delta
BenchmarkS3HeadObject-4     511            444            -13.11%

benchmark                   old bytes     new bytes     delta
BenchmarkS3HeadObject-4     68694         61968         -9.79%
```

```go
package main

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/aws/aws-sdk-go/aws"
	"github.com/aws/aws-sdk-go/aws/credentials"
	"github.com/aws/aws-sdk-go/aws/endpoints"
	"github.com/aws/aws-sdk-go/aws/session"
	"github.com/aws/aws-sdk-go/service/s3"
)

func BenchmarkS3HeadObject(b *testing.B) {
	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(http.StatusOK)
	}))

	sess := session.Must(session.NewSession(&aws.Config{
		Credentials:      credentials.NewStaticCredentials("Key", "Secret", "Token"),
		Endpoint:         aws.String(server.URL),
		S3ForcePathStyle: aws.Bool(true),
		DisableSSL:       aws.Bool(true),
		Region:           aws.String(endpoints.UsWest2RegionID),
	}))
	svc := s3.New(sess)
	b.ResetTimer()

	for i := 0; i < b.N; i++ {
		req, _ := svc.HeadObjectRequest(&s3.HeadObjectInput{
			Bucket:    aws.String("somebucketname"),
			Key:       aws.String("keyname"),
			VersionId: aws.String("someVersion"),
			IfMatch:   aws.String("IfMatch"),
		})
		if err := req.Send(); err != nil {
			b.Fatal(err)
		}
	}
}

```